Fix/transport times

+++ src/Modules/Tours/Explorer.Tours.Core/UseCases/Administration/TourService.cs
@@ -11,14 +11,12 @@ namespace Explorer.Tours.Core.UseCases.Administration;
 public class TourService : ITourService
 {
     private readonly ITourRepository _tourRepository;
-    private readonly ITransportTimeRepository _timeRepository;
     private readonly IEquipmentRepository _equipmentRepository;
     private readonly IMapper _mapper;
 
-    public TourService(ITourRepository tourRepository, ITransportTimeRepository timeRepository, IEquipmentRepository equipmentRepository, IMapper mapper)
+    public TourService(ITourRepository tourRepository, IEquipmentRepository equipmentRepository, IMapper mapper)
     {
         _tourRepository = tourRepository;
-        _timeRepository = timeRepository;
         _equipmentRepository = equipmentRepository;
         _mapper = mapper;
     }
@@ -26,7 +24,7 @@ public class TourService : ITourService
     public bool Archive(long id)
     {
         var tour = GetById(id);
-        if(tour == null) return false;
+        if (tour == null) return false;
 
         if (tour.Status == TourStatusDto.Archived)
         {
@@ -51,6 +49,9 @@ public class TourService : ITourService
     public void Delete(long id, long authorId)
     {
         var tour = _tourRepository.Get(id);
+        if (tour == null)
+            throw new KeyNotFoundException($"Tour with id {id} not found.");
+
         if (tour.CreatorId != authorId)
             throw new InvalidOperationException("Can't delete someone else's tour");
         _tourRepository.Delete(id);
@@ -86,10 +87,6 @@ public class TourService : ITourService
     public bool Publish(long id)
     {
         Tour? tourToUpdate = _tourRepository.Get(id);
-        // No logic implemented yet
-        //Additional validation for transport time
-        //List<TransportTime> transportTimes = _timeRepository.GetByTourId(id).ToList();
-        //if (tourToUpdate != null && transportTimes.Count >= 2)
         if (tourToUpdate != null)
         {
             if (tourToUpdate.Publish())
@@ -158,7 +155,7 @@ public class TourService : ITourService
     {
         var tour = _tourRepository.Get(id);
         var equip = _equipmentRepository.Get(equipmentId);
-        if(tour.CreatorId != authorId)
+        if (tour.CreatorId != authorId)
             throw new InvalidOperationException("Can't add equipment to someone else's tour");
         tour.AddEquipment(equip);
         var result = _tourRepository.Update(tour);
@@ -168,10 +165,46 @@ public class TourService : ITourService
     {
         var tour = _tourRepository.Get(id);
         var equip = _equipmentRepository.Get(equipmentId);
-        if(tour.CreatorId != authorId)
+        if (tour.CreatorId != authorId)
             throw new InvalidOperationException("Can't remove equipment from someone else's tour");
         tour.RemoveEquipment(equip);
         var result = _tourRepository.Update(tour);
         return _mapper.Map<TourDto>(result);
+    }
+
+    public TransportTimeDto AddTransportTime(long tourId, TransportTimeDto timeDto, long authorId)
+    {
+        if (timeDto.Type == TransportTypeDto.Unknown)
+            throw new ArgumentException("Transport type cannot be unknown.");
+            
+        var tour = _tourRepository.Get(tourId);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't add transport time to someone else's tour");
+        var transportTime = tour.AddTransportTime(_mapper.Map<TransportTime>(timeDto));
+        _tourRepository.Update(tour);
+
+        return _mapper.Map<TransportTimeDto>(transportTime);
+    }
+    public TransportTimeDto UpdateTransportTime(long tourId, TransportTimeDto timeDto, long authorId)
+    {
+        if (timeDto.Type == TransportTypeDto.Unknown)
+            throw new ArgumentException("Transport type cannot be unknown.");
+
+        var tour = _tourRepository.Get(tourId);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't update transport time from someone else's tour");
+        var tt = tour.UpdateTransportTime(_mapper.Map<TransportTime>(timeDto));
+        _tourRepository.Update(tour);
+
+        return _mapper.Map<TransportTimeDto>(tt);
+    }
+
+    public void DeleteTransportTime(long tourId, long timeId, long authorId)
+    {
+        var tour = _tourRepository.Get(tourId);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't delete transport time from someone else's tour");
+        tour.DeleteTransportTime(timeId);
+        var result = _tourRepository.Update(tour);
     }
 }
